re-arrange the plenaries to allow for namespaced profiles
Namespaced profiles are profiles in sub-directories. The
code now supports namespaced profiles and it makes things simpler,
so that's good. The build.xml now just builds "objects" (everything
under the cfg directory and doesn't have to care about hosts vs
clusters. Hosts can be either namespaced, or not namespaced, or
both. This is to allow for transition from our current model
(non-namespaced) to a new namespaced model, with a transition
in-between that is all config driven. Once we've transitioned to
the new model, we can drop some unneccessary patches from AII
that attempts client-side namespacing.

+++ lib/python2.5/aquilon/server/templates/base.py
@@ -145,10 +145,11 @@ class Plenary(object):
         """
 
         self.remove(None, locked)
-        qdir = self.config.get("broker", "quattordir")
-        xmlfile = os.path.join(qdir, "build", "xml", domain, name+".xml")
-        remove_file(xmlfile)
-        depfile = os.path.join(qdir, "build", "xml", domain, name+".xml.dep")
-        remove_file(depfile)
+        if self.template_type == "object":
+            qdir = self.config.get("broker", "quattordir")
+            xmlfile = os.path.join(qdir, "build", "xml", domain, name+".xml")
+            remove_file(xmlfile)
+            depfile = os.path.join(qdir, "build", "xml", domain, name+".xml.dep")
+            remove_file(depfile)
